+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
   pattern.
 - Field `successful_on` added to the `http` processor.
 - The `statsd` metrics type now supports Datadog or InfluxDB tagging.
+- Field `sync_response.headers` added to `http_server` input.
 
 ### Changed